refactor(query): merge cache and cache_manager
