+++ src/query/storages/common/cache/Cargo.toml
@@ -11,16 +11,22 @@ doctest = false
 test = true
 
 [dependencies]
+arrow = { workspace = true }
+databend-common-arrow = { workspace = true }
+databend-common-base = { workspace = true }
+databend-common-cache = { workspace = true }
+databend-common-catalog = { workspace = true }
+databend-common-config = { workspace = true }
+databend-common-exception = { workspace = true }
+databend-common-metrics = { workspace = true }
+databend-storages-common-index = { workspace = true }
+databend-storages-common-table-meta = { workspace = true }
+
 async-backtrace = { workspace = true }
 async-trait = { workspace = true }
 bytes = { workspace = true }
 crc32fast = "1.3.2"
 crossbeam-channel = "0.5.6"
-databend-common-base = { workspace = true }
-databend-common-cache = { workspace = true }
-databend-common-config = { workspace = true }
-databend-common-exception = { workspace = true }
-databend-common-metrics = { workspace = true }
 hex = "0.4.3"
 log = { workspace = true }
 parking_lot = { workspace = true }
