+++ src/query/storages/common/cache/src/lib.rs
@@ -16,12 +16,19 @@
 #![feature(associated_type_defaults)]
 
 mod cache;
+mod caches;
+mod manager;
 mod providers;
 mod read;
 
 pub use cache::CacheAccessor;
 pub use cache::Unit;
+pub use caches::BlockMetaCache;
+pub use caches::CachedObject;
+pub use caches::MemSizedMeter;
+pub use caches::SizedColumnArray;
 pub use databend_common_cache::CountableMeter;
+pub use manager::CacheManager;
 pub use providers::DiskCacheError;
 pub use providers::DiskCacheKey;
 pub use providers::DiskCacheResult;
