--- src/query/storages/common/cache_manager/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "databend-storages-common-cache-manager"
-version = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-publish = { workspace = true }
-edition = { workspace = true }
-
-[features]
-
-[dependencies]
-arrow = { workspace = true }
-databend-common-arrow = { workspace = true }
-databend-common-base = { workspace = true }
-databend-common-cache = { workspace = true }
-databend-common-catalog = { workspace = true }
-databend-common-config = { workspace = true }
-databend-common-exception = { workspace = true }
-databend-storages-common-cache = { workspace = true }
-databend-storages-common-index = { workspace = true }
-databend-storages-common-table-meta = { workspace = true }
-log = { workspace = true }
-
-[lints]
-workspace = true
