+++ src/query/storages/fuse/Cargo.toml
@@ -43,7 +43,6 @@ databend-common-storage = { workspace = true }
 databend-enterprise-fail-safe = { workspace = true }
 databend-storages-common-blocks = { workspace = true }
 databend-storages-common-cache = { workspace = true }
-databend-storages-common-cache-manager = { workspace = true }
 databend-storages-common-index = { workspace = true }
 databend-storages-common-io = { workspace = true }
 databend-storages-common-pruner = { workspace = true }
