+++ src/query/storages/fuse/src/io/read/block/block_reader_merge_io.rs
@@ -21,9 +21,9 @@ use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::ColumnId;
 use databend_storages_common_cache::CacheAccessor;
+use databend_storages_common_cache::SizedColumnArray;
 use databend_storages_common_cache::TableDataCache;
 use databend_storages_common_cache::TableDataCacheKey;
-use databend_storages_common_cache_manager::SizedColumnArray;
 use enum_as_inner::EnumAsInner;
 
 type ChunkIndex = usize;
