+++ src/query/storages/fuse/src/io/read/block/block_reader_merge_io_async.rs
@@ -24,8 +24,8 @@ use databend_common_exception::Result;
 use databend_common_expression::ColumnId;
 use databend_common_metrics::storage::*;
 use databend_storages_common_cache::CacheAccessor;
+use databend_storages_common_cache::CacheManager;
 use databend_storages_common_cache::TableDataCacheKey;
-use databend_storages_common_cache_manager::CacheManager;
 use databend_storages_common_table_meta::meta::ColumnMeta;
 use futures::future::try_join_all;
 use opendal::Operator;
