+++ src/query/storages/fuse/src/io/read/bloom/column_filter_reader.rs
@@ -19,10 +19,11 @@ use databend_common_expression::Column;
 use databend_common_expression::ColumnId;
 use databend_common_metrics::storage::metrics_inc_block_index_read_bytes;
 use databend_storages_common_cache::CacheKey;
+use databend_storages_common_cache::CachedObject;
 use databend_storages_common_cache::InMemoryCacheReader;
 use databend_storages_common_cache::LoadParams;
 use databend_storages_common_cache::Loader;
-use databend_storages_common_cache_manager::{CachedObject, MemSizedMeter};
+use databend_storages_common_cache::MemSizedMeter;
 use databend_storages_common_index::filters::Filter;
 use databend_storages_common_index::filters::Xor8Filter;
 use databend_storages_common_table_meta::meta::SingleColumnMeta;
