+++ src/query/storages/fuse/src/operations/gc.rs
@@ -25,8 +25,8 @@ use databend_common_exception::Result;
 use databend_common_meta_app::schema::ListIndexesByIdReq;
 use databend_common_meta_app::schema::TableIndex;
 use databend_storages_common_cache::CacheAccessor;
+use databend_storages_common_cache::CachedObject;
 use databend_storages_common_cache::LoadParams;
-use databend_storages_common_cache_manager::CachedObject;
 use databend_storages_common_index::BloomIndexMeta;
 use databend_storages_common_index::InvertedIndexFile;
 use databend_storages_common_index::InvertedIndexMeta;
