+++ src/query/storages/fuse/src/pruning/fuse_pruner.rs
@@ -27,9 +27,9 @@ use databend_common_expression::SEGMENT_NAME_COL_NAME;
 use databend_common_functions::BUILTIN_FUNCTIONS;
 use databend_common_sql::field_default_value;
 use databend_common_sql::BloomIndexColumns;
+use databend_storages_common_cache::BlockMetaCache;
 use databend_storages_common_cache::CacheAccessor;
-use databend_storages_common_cache_manager::BlockMetaCache;
-use databend_storages_common_cache_manager::CacheManager;
+use databend_storages_common_cache::CacheManager;
 use databend_storages_common_index::RangeIndex;
 use databend_storages_common_pruner::BlockMetaIndex;
 use databend_storages_common_pruner::InternalColumnPruner;
