+++ src/query/storages/system/Cargo.toml
@@ -37,7 +37,6 @@ databend-common-storages-stream = { workspace = true }
 databend-common-storages-view = { workspace = true }
 databend-common-users = { workspace = true }
 databend-storages-common-cache = { workspace = true }
-databend-storages-common-cache-manager = { workspace = true }
 futures = { workspace = true }
 itertools = { workspace = true }
 jsonb = { workspace = true }
